requirements.txt
streamlit==1.38.0
librosa==0.10.2
soundfile==0.12.1
pydub==0.25.1
numpy==1.26.4
matplotlib==3.9.2
python-dotenv==1.0.1
openai==1.45.0
pydub==0.25.1